search_backend/requirements.txt
elasticsearch==7.16.3
Flask==2.0.3
Flask_RESTful==0.3.9
flask_apispec==0.11.0
flask-cors==3.0.10
pydantic==1.9.0
pyyaml==6.0
uWSGI==2.0.20; platform_system != "Windows"